Module 6 Critical Thinking
This week’s discussion has provided an introduction to Lightning and Shading, particularly with respect to reflection and Vector Computations. In this assignment, you will briefly discuss a sphere object (in one paragraph), which is used as an example curved surface to illustrate shading calculations. Then use the code snippet starting in section 6.6 to design an interactive approximated and recursively subdivided sphere, using WebGL. Finally, complete a two paragraph reflection on what you did and how it went.